refactor(social-messenger-ts): organize app modules
