--- social-messenger-ts/src/components/MessagingInput/index.ts
@@ -1,1 +1,0 @@
-export * from './SendButton';
