+++ social-messenger-ts/src/components/SendButton/SendButton.tsx
@@ -2,7 +2,7 @@ import { SendIcon } from '../../assets';
 import React from 'react';
 import type { SendButtonProps } from 'stream-chat-react';
 
-export const SendButton = ({sendMessage, ...rest}: SendButtonProps) => (
+const SendButton = ({sendMessage, ...rest}: SendButtonProps) => (
   <button
     className='str-chat__send-button'
     onClick={sendMessage}
@@ -11,4 +11,6 @@ export const SendButton = ({sendMessage, ...rest}: SendButtonProps) => (
   >
     <SendIcon />
   </button>
-)
+);
+
+export default SendButton;
