+++ social-messenger-ts/src/components/index.ts
@@ -1,9 +1,11 @@
 export { default as AvatarGroup } from './AvatarGroup/AvatarGroup';
+export { default as ChannelInner } from './ChannelInner/ChannelInner';
 export { default as CreateChannel } from './CreateChannel/CreateChannel';
 export { default as MessagingChannelHeader } from './MessagingChannelHeader/MessagingChannelHeader';
 export { default as MessagingChannelList } from './MessagingChannelList/MessagingChannelList';
 export { default as MessagingChannelListHeader } from './MessagingChannelList/MessagingChannelListHeader';
 export { default as MessagingChannelPreview } from './MessagingChannelPreview/MessagingChannelPreview';
-export * from './MessagingInput';
+export { default as MessagingSidebar } from './MessagingSidebar/MessagingSidebar';
 export { default as MessagingThreadHeader } from './MessagingThreadHeader/MessagingThreadHeader';
+export {default as SendButton} from './SendButton/SendButton';
 export { default as WindowControls } from './WindowControls/WindowControls';
